Consolidate project into single-page assets
