--- data/README.md
@@ -1,15 +1,0 @@
-# Dataset Sources
-
-This directory contains small reference datasets used by the Actuarial Toolkit. Each JSON file now records its origin in a `source` field (no hyperlinks).
-
-| File | Description | Source |
-| --- | --- | --- |
-| `lifetable_US_2022.json` | U.S. period life table for 2022 | Arias E, Tejada-Vera B. *United States Life Tables, 2022.* National Vital Statistics Reports; 73(1). 2024.  |
-| `hr_physical_activity.json` | Hazard ratios for leisure-time physical activity (MET-hours/week) | Arem H et al. *Leisure Time Physical Activity and Mortality: A Dose-Response Analysis.* JAMA Intern Med. 2015.  |
-| `hr_bmi.json` | Hazard ratios for body-mass index categories | Global BMI Mortality Collaboration. *Body-mass index and all-cause mortality.* Lancet. 2016.  |
-| `hr_smoking.json` | Hazard ratios for current and former smoking | Jha P et al. *21st-Century Hazards of Smoking and Benefits of Cessation in the United States.* N Engl J Med. 2013.  |
-| `hr_alcohol.json` | Hazard ratios for alcohol consumption (drinks per day) | Wood AM et al. *Risk thresholds for alcohol consumption: combined analysis of individual-participant data for 599 912 current drinkers in 83 prospective studies.* Lancet. 2018.  |
-| `screening_crc.json` | Life-years and quality-adjusted life-years gained from colorectal cancer screening | Knudsen AB et al. *Colorectal Cancer Screening: An Updated Modeling Study for the U.S. Preventive Services Task Force.* JAMA. 2021.  |
-| `screening_breast.json` | Life-years and quality-adjusted life-years gained from breast cancer screening | U.S. Preventive Services Task Force. *Screening for Breast Cancer: Recommendation Statement.* Ann Intern Med. 2016.  |
-| `hale_weights.json` | Age-specific weights for health-adjusted life expectancy | Institute for Health Metrics and Evaluation. *Global Burden of Disease Study 2019 (HALE).*  |
-
