--- data/hale_weights.json
@@ -1,9 +1,0 @@
-{
-  "F": {
-    "0": 0.98, "10": 0.97, "20": 0.96, "30": 0.95, "40": 0.93, "50": 0.90, "60": 0.85, "70": 0.78, "80": 0.68, "90": 0.50, "100": 0.30
-  },
-  "M": {
-    "0": 0.98, "10": 0.97, "20": 0.96, "30": 0.95, "40": 0.92, "50": 0.88, "60": 0.82, "70": 0.74, "80": 0.63, "90": 0.45, "100": 0.28
-  },
-  "source": "IHME. Global Burden of Disease Study 2019 (HALE)"
-}
