--- data/hr_alcohol.json
@@ -1,12 +1,0 @@
-{
-  "gramsPerDrink": 14,
-  "dose_HR": [
-    { "drinksPerDay": 0, "HR": 1.00 },
-    { "drinksPerDay": 1, "HR": 1.03 },
-    { "drinksPerDay": 2, "HR": 1.06 },
-    { "drinksPerDay": 3, "HR": 1.12 },
-    { "drinksPerDay": 4, "HR": 1.20 },
-    { "drinksPerDay": 6, "HR": 1.40 }
-  ],
-  "source": "Wood AM et al. Risk thresholds for alcohol consumption. Lancet 2018"
-}
