--- data/hr_physical_activity.json
@@ -1,5 +1,0 @@
-{
-  "bins_MET_h_week": [0, 7.5, 15, 22.5, 37.5, 60, 90],
-  "HR":              [1.00, 0.81, 0.69, 0.61, 0.55, 0.53, 0.53],
-  "source": "Arem H et al. Leisure time physical activity and mortality. JAMA Intern Med 2015"
-}
