--- data/lifetable_US_2022.json
@@ -1,13 +1,0 @@
-{
-  "F": { "qx": {
-    "0": 0.00527, "1": 0.00040, "5": 0.00012, "10": 0.00008, "15": 0.00029, "20": 0.00052, "25": 0.00062, "30": 0.00074,
-    "35": 0.00102, "40": 0.00142, "45": 0.00206, "50": 0.00305, "55": 0.00453, "60": 0.00677, "65": 0.01010, "70": 0.01547,
-    "75": 0.02447, "80": 0.03967, "85": 0.06540, "90": 0.10800, "95": 0.17600, "100": 0.28000, "105": 0.45000, "110": 1.0
-  }},
-  "M": { "qx": {
-    "0": 0.00559, "1": 0.00045, "5": 0.00013, "10": 0.00009, "15": 0.00037, "20": 0.00079, "25": 0.00096, "30": 0.00116,
-    "35": 0.00157, "40": 0.00220, "45": 0.00317, "50": 0.00463, "55": 0.00693, "60": 0.01048, "65": 0.01580, "70": 0.02405,
-    "75": 0.03733, "80": 0.05941, "85": 0.09502, "90": 0.15000, "95": 0.23000, "100": 0.35000, "105": 0.52000, "110": 1.0
-  }},
-  "source": "Arias E, Tejada-Vera B. United States Life Tables, 2022. National Vital Statistics Reports; vol 73 no 1. 2024"
-}
