--- data/screening_breast.json
@@ -1,6 +1,0 @@
-{
-  "default_schedule": { "startAge": 40, "stopAge": 74, "intervalYears": 2 },
-  "LYG_per_1000": 60,
-  "QALY_per_1000": 45,
-  "source": "U.S. Preventive Services Task Force. Screening for Breast Cancer: Recommendation Statement. Ann Intern Med. 2016"
-}
