+++ index.html
@@ -25,7 +25,6 @@
     }
   </script>
   <script src="https://cdn.tailwindcss.com"></script>
-  <link rel="stylesheet" href="./theme/style.css" />
   <link rel="stylesheet" href="./style.css" />
 </head>
 <body class="bg-slate-50 text-slate-900 antialiased">
@@ -336,6 +335,6 @@
     </div>
   </noscript>
 
-  <script type="module" src="js/app.js"></script>
+  <script src="script.js"></script>
 </body>
 </html>
